chore: reorganize components and update imports
