+++ apps/expo/src/forms/SignupForm.tsx
@@ -14,7 +14,7 @@ import useSignin from "../hooks/useSignin";
 
 import type { UserSignup } from "@acme/schema/src/types";
 import type { FC } from "react";
-import { AppButton } from "../components/AppButton";
+import { AppButton } from "../components/buttons/AppButton";
 
 interface Props {
   onSubmit: (data: UserSignup) => void;
