+++ apps/expo/src/screens/create-collection.tsx
@@ -7,7 +7,7 @@ import {
   TextInput,
   TouchableOpacity,
 } from "react-native";
-import { AppButton } from "../components/AppButton";
+import { AppButton } from "../components/buttons/AppButton";
 import { FontAwesome } from "@expo/vector-icons";
 import { MaterialIcons } from "@expo/vector-icons";
 import SelectDropdown from "react-native-select-dropdown";
