+++ apps/expo/src/screens/profile.tsx
@@ -3,8 +3,8 @@ import { SafeAreaView, View, TouchableOpacity, Text } from "react-native";
 import ProfileHeader from "../components/headers/ProfileHeader";
 import ProfileDetailsSection from "../components/profile-details/details-section/ProfileDetailsSection";
 import RectangleProfileIcon from "../icons/RectangleProfileIcon";
-import { CollectionsTab } from "../components/Collections";
-import { TabContent } from "../components/TabContent";
+import { CollectionsTab } from "../components/my-collections/Collections";
+import { TabContent } from "../components/my-library/TabContent";
 import { AboutUser } from "../components/AboutUser";
 import Footer from "../components/Footer";
 import { trpc } from "../utils/trpc";
